Extract parser statement dispatch
