+++ crates/truthdb-sql/src/parser/session.rs
@@ -1,3 +1,4 @@
+use super::token::is_reserved;
 use super::*;
 
 impl Parser {
